Delete option is implemented

+++ BloggingProject/src/main/java/com/blog/BloggingProject/repository/PostRepo.java
@@ -3,6 +3,5 @@ package com.blog.BloggingProject.repository;
 import com.blog.BloggingProject.model.Post;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PostRepo extends JpaRepository<Post,Integer> {
-
+public interface PostRepo extends JpaRepository<Post,Long> {
 }
